pyproject.toml
[project]
name = "lang-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.31",
    "langchain-openai>=0.3.2",
    "langgraph>=0.2.67",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "langgraph-cli[inmem]>=0.1.56",
    "langgraph-sdk>=0.1.51",
    "langsmith>=0.3.1",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
    "trustcall>=0.0.28",
    "wikipedia>=1.4.0",
]
